Remove "install(EXPORT ...)" from main CMakeLists.txt
CMake will not allow linking of a library witout install/export to a
library with it.

This is a problem when you need to link some low-level libraries (mbed
TLS, lwIP, RTOS, ...) to aziotsharedutil. These low-level libraries will
not have install/export configured (because it makes no sense), which
will cause a CMake error in such scenario.

https://github.com/Azure/azure-iot-sdk-c/issues/1188